Structure/validate (#282)
* Factor out validate utilities

* Factor out validateOnFollow

* Factor out validateContents

* Factor out validateSubtype

* Factor out validateTitle

* Factor out validateFollows

* Factor out validateParents

* Factor out validateAvatar

* Factor out validateMain

* Factor out validateDescription

* Factor out validateLinks

* Factor out validateType

* Factor out validateAuthors

* Factor out validateP2Pcommons

* Factor out validateUrl

* Factor out validateParentsOnUpdate

* Structure validation tests (#283)

* Factor out validateTitle tests

* Factor out validateDescription tests

* Factor out validateUrl tests

* Factor out validateLinks tests

* Factor out validateP2Pcommons tests

* Factor out validateType tests

* Factor out validateSubtype tests

* Factor out validateMain tests

* Factor out validateAvatar tests

* Factor out validateAuthors tests

* Factor out validateParents tests

* Factor out validateFollows tests

* Factor out validateContents tests

* Factor out validatePartial tests

* Factor out validateOnFollow tests

* Factor out validateOnRegister tests

* Update test to load from helper files

* Factor out complete validation tests
