+++ index.js
@@ -46,7 +46,7 @@ const {
   validateDescription,
   validateUrl,
   validateLinks,
-  validateP2pcommons,
+  validateP2Pcommons,
   validateType,
   validateSubtype,
   validateMain,
@@ -2470,7 +2470,7 @@ SDK.validations = {
   validateDescription,
   validateUrl,
   validateLinks,
-  validateP2pcommons,
+  validateP2Pcommons,
   validateType,
   validateSubtype,
   validateMain,
