+++ tests/example-key.js
@@ -1,0 +1,24 @@
+exports.exampleKey1 =
+  '4e01f6848573dcc0a712bd2482e6a3074310757448cd4a78fe219547fc2e484f'
+exports.exampleKey1V5 =
+  '4e01f6848573dcc0a712bd2482e6a3074310757448cd4a78fe219547fc2e484f+5'
+exports.exampleKey1V123 =
+  '4e01f6848573dcc0a712bd2482e6a3074310757448cd4a78fe219547fc2e484f+123'
+exports.exampleKey2 =
+  '8af39eb4a3eb3252141718f876d29220b8d6f539a045e833e9556aff2a5eacd8'
+exports.exampleKey2V5 =
+  '8af39eb4a3eb3252141718f876d29220b8d6f539a045e833e9556aff2a5eacd8+5'
+exports.exampleKey2V40 =
+  '8af39eb4a3eb3252141718f876d29220b8d6f539a045e833e9556aff2a5eacd8+40'
+exports.exampleKey2V123 =
+  '8af39eb4a3eb3252141718f876d29220b8d6f539a045e833e9556aff2a5eacd8+123'
+exports.exampleKey3 =
+  'cca6eb69a3ad6104ca31b9fee7832d74068db16ef2169eaaab5b48096e128342'
+exports.exampleKey3V5 =
+  'cca6eb69a3ad6104ca31b9fee7832d74068db16ef2169eaaab5b48096e128342+5'
+exports.exampleKey4 =
+  'f7daadc2d624df738abbccc9955714d94cef656406f2a850bfc499c2080627d4'
+exports.exampleKey5V12 =
+  'f0abcd6b1c4fc524e2d48da043b3d8399b96d9374d6606fca51182ee230b6b59+12'
+exports.exampleKey6V4032 =
+  '527f404aa77756b91cba4e3ba9fe30f72ee3eb5eef0f4da87172745f9389d1e5+4032'
